requirements.txt
wordcloud
pillow
numpy
stop_words
multidict
